Refactor EthContracts to typescript
