+++ libs/src/services/ethContracts/index.ts
@@ -1,0 +1,1 @@
+export * from './EthContracts'
