+++ libs/src/services/ethContracts/registryClient.ts
@@ -1,19 +1,19 @@
 import { Utils } from '../../utils'
-import type { Web3Manager } from '../web3Manager'
 import type { AbiItem } from 'web3-utils'
 import type Web3 from 'web3'
 import type { Contract } from 'web3-eth-contract'
+import type { EthWeb3Manager } from '../ethWeb3Manager'
 
 export class RegistryClient {
-  web3Manager: Web3Manager
-  contractABI: AbiItem
+  web3Manager: EthWeb3Manager
+  contractABI: AbiItem[]
   contractAddress: string
   web3: Web3
   Registry: Contract
 
   constructor(
-    web3Manager: Web3Manager,
-    contractABI: AbiItem,
+    web3Manager: EthWeb3Manager,
+    contractABI: AbiItem[],
     contractAddress: string
   ) {
     this.web3Manager = web3Manager
@@ -24,7 +24,7 @@ export class RegistryClient {
     this.Registry = new this.web3.eth.Contract(contractABI, contractAddress)
   }
 
-  async getContract(contractRegistryKey: string) {
+  async getContract(contractRegistryKey: string): Promise<string> {
     Utils.checkStrLen(contractRegistryKey, 32)
     return this.Registry.methods
       .getContract(Utils.utf8ToHex(contractRegistryKey))
